Ignore already killed processes

+++ src/index.ts
@@ -16,9 +16,10 @@ export async function endChildProcesses(): Promise<void> {
     try {
       process.kill(processID)
     } catch (e) {
-      // const err = e as Error
-      console.log(util.inspect(e), true, Infinity)
-      throw new Error(`cannot kill process ${processID} (${child.COMMAND}): ${(e as Error).message}`)
+      const err = e as Error
+      if (!err.message.includes("ESRCH")) {
+        throw new Error(`cannot kill process ${processID} (${child.COMMAND}): ${(e as Error).message}`)
+      }
     }
   }
   await delay(1)
